pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
  "src/mgost"
]

[project]
name="mgost"
version="1.2.1"
authors = [
  {name="Геворкян Артём", email="tima-1324@mail.ru"}
]
license = {text = "GNUv3"}
description = "Конвертер MD файлов в ГОСТ стандарт"
dependencies = [
  'python-docx >= 1.1.2',
  'pillow >= 10.4.0',
  'marko >= 2.1.2',
  'requests',
  'sympy >= 1.13.3',
  'latex2mathml',
]
readme = "readme.md"
classifiers = [
  "Development Status :: 1 - Planning"
  , "Development Status :: 4 - Beta"
  , "Operating System :: POSIX :: Linux"
  , "Operating System :: Microsoft :: Windows"
  , "Programming Language :: Python"
  , "Programming Language :: Python :: 3"
  , "Programming Language :: Python :: 3 :: Only"
  , "Programming Language :: Python :: 3.12"
  , "Natural Language :: Russian"
]
requires-python = ">= 3.12"

[project.urls]
homepage = "https://github.com/ArtichaTM/MarkdownGost"

[project.scripts]
mgost = "mgost.cli:main"

[project.optional-dependencies]
tests = [
  'flake8' # py3.9 7.0.0
]
build = [
  'build'
]